exception handling and session management complete

+++ src/main/java/com/ignou/shoppingportal/services/ProductService.java
@@ -1,7 +1,6 @@
 package com.ignou.shoppingportal.services;
 
 import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
 
 import javax.persistence.EntityManager;
@@ -10,16 +9,14 @@ import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.ignou.shoppingportal.controller.ProductController;
 import com.ignou.shoppingportal.model.Product;
 
 @Service
 public class ProductService {
 	@Autowired
 	EntityManager entityManager;
-	
+
 	private static Logger logger = Logger.getLogger(ProductService.class);
-	
 
 	public boolean addProduct(Product product) {
 		entityManager.getTransaction().begin();
@@ -41,33 +38,35 @@ public class ProductService {
 	}
 
 	public Product getProduct(int productId) {
-		Product product = null;
 		entityManager.getTransaction().begin();
-		product = entityManager.find(Product.class, productId);
-		  entityManager.getTransaction().commit();
+		Product product = entityManager.find(Product.class, productId);
+		if (product == null) {
+			throw new NullPointerException();
+		}
+		entityManager.getTransaction().commit();
 		return product;
 	}
 
 	public List<Product> allProduct() {
 		entityManager.getTransaction().begin();
-		
-		  List<Product> product=new ArrayList<Product>();
-		  logger.info("Return all product");
-		  product= entityManager.createQuery("select e from Product e").getResultList();
-		  entityManager.getTransaction().commit();
-		  return product;
-		
-		  
+		List<Product> product = new ArrayList<Product>();
+		logger.info("Return all product");
+		product = entityManager.createQuery("select e from Product e").getResultList();
+		entityManager.getTransaction().commit();
+		return product;
+
 	}
 
-	public boolean deleteProduct(int productId) {
-		Product product = null;
+	public Product deleteProduct(int productId) {
 
 		entityManager.getTransaction().begin();
-		product = entityManager.find(Product.class, productId);
+		Product product = entityManager.find(Product.class, productId);
+		if (product == null) {
+			throw new NullPointerException();
+		}
 		entityManager.remove(product);
 		entityManager.getTransaction().commit();
-		return true;
+		return product;
 	}
 
 }
